Rename AggregationConverter to AggregationContainerConverter; and update parsing code.
